+++ src/com/sugarglider/book/BookRunner.java
@@ -4,7 +4,7 @@ public class BookRunner {
     public static void main(String[] args) {
         Book artOfComputerProgramming = new Book();
 
-        artOfComputerProgramming.setNoOfCopies(5);
+        artOfComputerProgramming.setNoOfCopies((short)5);
         System.out.println(artOfComputerProgramming.getNoOfCopies());
         artOfComputerProgramming.openBook();
     }
